+++ flake.nix
@@ -70,7 +70,7 @@
             ++ [ spec-driver-pkg ];
 
           mcpJailOptions = with jailLib.combinators; [
-            (try-readwrite "/run/user/1000/satan-mcp-smoke.sock")
+            (try-readwrite "/run/user/1000/satan/mcp/mcp.sock")
             (try-fwd-env "XDG_RUNTIME_DIR")
             (try-fwd-env "SATAN_MCP_SOCKET")
             # this would be insecure (allow arbitrary elisp execution):